svdl: compute spectral gap more reasonably when l = 1
Fix #55

The code for estimating the spectral gap set the estimate to Inf and reduced it successively as various differences between successive singular values are computed. However, this initialization to Inf meant that when only one singular value is requested, the spectral gap is always calculated as Inf since no gap is actually computed.

Instead, if only one singular value is requested, then we simply use only the estimate that doesn't need an estimate for the spectral gap.

Also fix some bounds on the tests.